requirements.txt
requests
pathway
google-generativeai>=0.7.0
python-dotenv
streamlit
pandas
